Break utils into three modules, move config helper class into utils too. (196 241)
